pyproject.toml
[project]
name = "feeltalk"
version = "0.1.0"
description = "AI chatbot powered by MistralAI using mistral-small-latest"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.2",
    "mistralai>=2.2.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
